# service/openapi/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Atlas Enterprise Agent API
  version: 0.1.0
  description: 单企业风险排查与 DOCX 报告生成 V1 API
servers:
  - url: http://localhost:8080
paths:
...
  /api/agent/messages:
    post:
      summary: Interpret one controlled natural-language task message
      operationId: sendAgentMessage
      description: >
        Creates a risk report update task or reads an existing task. The
        endpoint never confirms evidence, adjusts a score, or generates a
        formal report from ambiguous natural language.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/OperatorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentMessageRequest'
      responses:
        '200':
          description: Parsed message, assistant response and optional task workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentMessageResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '403':
          description: Task does not belong to the declared operator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/agent/conversations:
    get:
      summary: List the current operator's conversations
      operationId: listAgentConversations
      parameters:
        - $ref: '#/components/parameters/OperatorId'
      responses:
        '200':
          description: Conversations ordered by most recently updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentConversation'
    post:
      summary: Create a durable Agent conversation
      operationId: createAgentConversation
      parameters:
        - $ref: '#/components/parameters/OperatorId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentConversationRequest'
      responses:
        '201':
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentConversation'
        '400':
          $ref: '#/components/responses/InvalidRequest'
  /api/agent/conversations/{conversationId}:
    get:
      summary: Read one conversation and its inherited task context
      operationId: getAgentConversation
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - $ref: '#/components/parameters/OperatorId'
      responses:
        '200':
          description: Conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentConversation'
        '403':
          description: Conversation belongs to another operator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/agent/conversations/{conversationId}/messages:
    get:
      summary: Read persisted messages in chronological order
      operationId: listAgentConversationMessages
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - $ref: '#/components/parameters/OperatorId'
      responses:
        '200':
          description: Persisted user and assistant messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentStoredMessage'
    post:
      summary: Send a message using durable multi-turn context
      operationId: sendAgentConversationMessage
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/OperatorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentMessageRequest'
      responses:
        '200':
          description: Persisted assistant response with current task workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentMessageResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '403':
          description: Conversation or task belongs to another operator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/tasks/{taskId}/risk-score/calculate:
    post:
      summary: 基于冻结快照和已确认事件计算风险分
      operationId: calculateRiskScore
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalculateRiskScoreRequest'
      responses:
        '200':
          description: 幂等评分快照
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskScore'
        '400':
          $ref: '#/components/responses/InvalidRequest'
  /api/tasks/{taskId}/risk-score/calculate-from-confirmed-evidence:
    post:
      summary: 基于运营人员已确认的公开证据计算风险分
      operationId: calculateRiskScoreFromConfirmedEvidence
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 幂等评分快照
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskScore'
        '400':
          $ref: '#/components/responses/InvalidRequest'
...
  /api/tasks/{taskId}/company-aliases:
    get:
      summary: 查询企业全称、简称、曾用名、品牌、门店等已登记身份词
      operationId: listCompanyAliases
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 企业身份词及主体关系
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompanyAlias'
    post:
      summary: 运营人员确认并补录企业身份词
      operationId: addCompanyAlias
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyAliasRequest'
      responses:
        '201':
          description: 已确认的企业身份词
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyAlias'
        '400':
          $ref: '#/components/responses/InvalidRequest'
...
  /api/tasks/{taskId}/risk-score:
    get:
      summary: 查询任务最新评分快照
      operationId: getLatestRiskScore
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 最新评分快照
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskScore'
        '404':
          description: 尚未计算评分
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/tasks/{taskId}/risk-score/{scoreSnapshotId}/adjustments:
    post:
      summary: 人工调整风险分并记录原因
      operationId: adjustRiskScore
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/ScoreSnapshotId'
        - $ref: '#/components/parameters/OperatorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdjustRiskScoreRequest'
      responses:
        '200':
          description: 调整后的分数、决定记录和最低分覆盖警告
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskScoreAdjustment'
        '400':
          $ref: '#/components/responses/InvalidRequest'
  /api/tasks/{taskId}/risk-score/history:
    get:
      summary: 查询任务全部版本化研判结果
      operationId: getRiskAssessmentHistory
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 按版本倒序返回系统研判和人工调整记录
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RiskAssessmentRevision'
  /api/tasks/{taskId}/risk-score/decisions:
    get:
      summary: 查询评分人工调整历史
      operationId: getRiskScoreDecisions
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 按时间排序的操作记录
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OperatorDecision'
...
components:
  parameters:
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
    OperatorId:
      name: X-Operator-Id
      in: header
      required: false
      schema:
        type: string
        default: local-operator
...
    ScoreSnapshotId:
      name: scoreSnapshotId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    InvalidRequest:
      description: 请求不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    CreateAgentConversationRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type:
            - string
            - 'null'
          maxLength: 128
    AgentConversation:
      type: object
      required:
        - conversation_id
        - title
        - created_at
        - updated_at
      properties:
        conversation_id:
          type: string
          format: uuid
        title:
          type: string
        task_id:
          type:
            - string
            - 'null'
          format: uuid
        company_query:
          type:
            - string
            - 'null'
        previous_report_file_id:
          type:
            - string
            - 'null'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AgentStoredMessage:
      type: object
      required:
        - message_id
        - role
        - content
        - required_inputs
        - suggested_actions
        - created_at
      properties:
        message_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [USER, ASSISTANT]
        content:
          type: string
        response_type:
          type:
            - string
            - 'null'
          enum:
            - TASK_CREATED
            - TASK_STATUS
            - NEEDS_INPUT
            - UNSUPPORTED_SCOPE
            - null
        parsed_intent:
          type:
            - string
            - 'null'
          enum:
            - CREATE_RISK_REPORT_TASK
            - QUERY_TASK_STATUS
            - UNSUPPORTED_SCOPE
            - UNKNOWN
            - null
        company_query:
          type:
            - string
            - 'null'
        task_id:
          type:
            - string
            - 'null'
          format: uuid
        required_inputs:
          type: array
          items:
            $ref: '#/components/schemas/AgentRequiredInput'
        suggested_actions:
          type: array
          items:
            $ref: '#/components/schemas/AgentSuggestedAction'
        created_at:
          type: string
          format: date-time
    AgentMessageRequest:
      type: object
      additionalProperties: false
      required:
        - message
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 4000
        task_id:
          type:
            - string
            - 'null'
          format: uuid
        previous_report_file_id:
          type:
            - string
            - 'null'
          maxLength: 256
          deprecated: true
          description: Historical compatibility field; ignored by new risk investigation tasks.
    AgentMessageResponse:
      type: object
      required:
        - conversation_id
        - message_id
        - type
        - assistant_message
        - parsed_intent
        - required_inputs
        - suggested_actions
      properties:
        conversation_id:
          type: string
          format: uuid
        message_id:
          type: string
          format: uuid
        type:
          type: string
          enum:
            - TASK_CREATED
            - TASK_STATUS
            - NEEDS_INPUT
            - UNSUPPORTED_SCOPE
        assistant_message:
          type: string
        parsed_intent:
          type: string
          enum:
            - CREATE_RISK_REPORT_TASK
            - QUERY_TASK_STATUS
            - UNSUPPORTED_SCOPE
            - UNKNOWN
        company_query:
          type:
            - string
            - 'null'
        workspace:
          anyOf:
            - $ref: '#/components/schemas/TaskWorkspace'
            - type: 'null'
        required_inputs:
          type: array
          items:
            $ref: '#/components/schemas/AgentRequiredInput'
        suggested_actions:
          type: array
          items:
            $ref: '#/components/schemas/AgentSuggestedAction'
    AgentRequiredInput:
      type: object
      required:
        - code
        - label
        - required
      properties:
        code:
          type: string
          enum:
            - COMPANY_QUERY
            - TASK_REFERENCE
        label:
          type: string
        required:
          type: boolean
    AgentSuggestedAction:
      type: object
      required:
        - code
        - label
        - method
        - endpoint
      properties:
        code:
          type: string
        label:
          type: string
        method:
          type: string
          enum: [GET, POST]
        endpoint:
          type: string
...
    CalculateRiskScoreRequest:
      type: object
      additionalProperties: false
      properties:
        confirmed_events:
          type: array
          default: []
          items:
            $ref: '#/components/schemas/ConfirmedRiskEvent'
    ConfirmedRiskEvent:
      type: object
      additionalProperties: false
      required:
        - risk_type
        - reference_id
      properties:
        risk_type:
          type: string
          description: 只有已确认的失联、欠薪或门店关闭事件会触发 V1 最低分
          enum: [OUT_OF_CONTACT, WAGE_ARREARS, STORE_CLOSURE]
        reference_id:
          type: string
        title:
          type:
            - string
            - 'null'
        evidence_ids:
          type: array
          default: []
          items:
            type: string
...
    CompanyAliasRequest:
      type: object
      additionalProperties: false
      required: [alias_name, alias_type, relation, source_evidence]
      properties:
        alias_name:
          type: string
          minLength: 1
          maxLength: 256
        alias_type:
          type: string
          enum: [SHORT_NAME, FORMER_NAME, BRAND, STORE, WEBSITE, SOCIAL_ACCOUNT]
        relation:
          type: string
          enum: [SAME_LEGAL_ENTITY, FORMER_IDENTITY, OWNED_BRAND, OPERATED_STORE, FRANCHISE_STORE, OTHER]
        source_evidence:
          type: string
          minLength: 1
          maxLength: 1000
    CompanyAlias:
      type: object
      required:
        - alias_id
        - atlas_company_id
        - alias_name
        - alias_type
        - relation
        - verification_status
        - source_system
        - created_at
        - updated_at
      properties:
        alias_id:
          type: string
          format: uuid
        atlas_company_id:
          type: string
          format: uuid
        alias_name:
          type: string
        alias_type:
          type: string
          enum: [LEGAL_NAME, SHORT_NAME, FORMER_NAME, BRAND, STORE, WEBSITE, SOCIAL_ACCOUNT]
        relation:
          type: string
          enum: [SAME_LEGAL_ENTITY, FORMER_IDENTITY, OWNED_BRAND, OPERATED_STORE, FRANCHISE_STORE, OTHER]
        verification_status:
          type: string
          enum: [CONFIRMED, REJECTED]
        source_system:
          type: string
        source_record_id:
          type: [string, 'null']
        source_evidence:
          type: [string, 'null']
        created_by:
          type: [string, 'null']
        valid_from:
          type: [string, 'null']
          format: date-time
        valid_to:
          type: [string, 'null']
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    RiskScore:
      type: object
      required:
        - score_snapshot_id
        - task_id
        - data_snapshot_id
        - rule_calculated_score
        - event_floor_score
        - original_score
        - manual_score
        - original_risk_level
        - manual_risk_level
        - rule_version
        - engine_version
        - input_hash
        - rule_hits
        - calculated_at
      properties:
        score_snapshot_id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        data_snapshot_id:
          type: string
          format: uuid
        legacy_score:
          type:
            - number
            - 'null'
          minimum: 0
          maximum: 10
        rule_calculated_score:
          type: number
          minimum: 0
          maximum: 10
        event_floor_score:
          type: number
          minimum: 0
          maximum: 10
        original_score:
          type: number
          minimum: 0
          maximum: 10
        manual_score:
          type: number
          minimum: 0
          maximum: 10
        original_risk_level:
          $ref: '#/components/schemas/RiskLevel'
        manual_risk_level:
          $ref: '#/components/schemas/RiskLevel'
        rule_version:
          type: string
        engine_version:
          type: string
        input_hash:
          type: string
        rule_hits:
          type: array
          items:
            $ref: '#/components/schemas/RiskRuleHit'
        calculated_at:
          type: string
          format: date-time
    RiskLevel:
      type: string
      enum: [LOW, MEDIUM_LOW, MEDIUM, MEDIUM_HIGH, HIGH]
    RiskRuleHit:
      type: object
      required:
        - rule_code
        - rule_name
        - score
        - score_role
        - references
      properties:
        rule_code:
          type: string
        rule_name:
          type: string
        risk_type:
          type:
            - string
            - 'null'
        score:
          type: number
        score_role:
          type: string
          enum: [RULE_SCORE, EVENT_FLOOR]
        references:
          type: array
          items:
            type: string
...
    RiskAssessmentRevision:
      type: object
      required:
        - assessment_revision_id
        - task_id
        - score_snapshot_id
        - data_snapshot_id
        - revision_no
        - trigger_type
        - rule_calculated_score
        - event_floor_score
        - original_score
        - final_score
        - original_risk_level
        - final_risk_level
        - rule_version
        - engine_version
        - source_labels
        - rule_labels
        - model_labels
        - final_labels
        - actor_type
        - actor_id
        - reason_code
        - reason_text
        - created_at
      properties:
        assessment_revision_id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        score_snapshot_id:
          type: string
          format: uuid
        data_snapshot_id:
          type: string
          format: uuid
        revision_no:
          type: integer
          minimum: 1
        trigger_type:
          type: string
          enum: [SYSTEM_CALCULATION, MANUAL_SCORE_ADJUSTMENT]
        legacy_score:
          type: [number, 'null']
        rule_calculated_score:
          type: number
        event_floor_score:
          type: number
        original_score:
          type: number
        final_score:
          type: number
        original_risk_level:
          $ref: '#/components/schemas/RiskLevel'
        final_risk_level:
          $ref: '#/components/schemas/RiskLevel'
        rule_version:
          type: string
        engine_version:
          type: string
        source_labels:
          type: array
          items:
            $ref: '#/components/schemas/RiskAssessmentLabel'
        rule_labels:
          type: array
          items:
            $ref: '#/components/schemas/RiskAssessmentLabel'
        model_labels:
          type: array
          items:
            $ref: '#/components/schemas/RiskAssessmentLabel'
        final_labels:
          type: array
          items:
            $ref: '#/components/schemas/RiskAssessmentLabel'
        actor_type:
          type: string
        actor_id:
          type: string
        reason_code:
          type: string
        reason_text:
          type: string
        created_at:
          type: string
          format: date-time
    AdjustRiskScoreRequest:
      type: object
      additionalProperties: false
      required:
        - manual_score
        - reason_code
        - reason_text
      properties:
        manual_score:
          type: number
          minimum: 0
          maximum: 10
        reason_code:
          type: string
          enum:
            - EVIDENCE_CORRECTION
            - ENTITY_MISMATCH
            - EVENT_RESOLVED
            - ADDITIONAL_CONTEXT
            - RULE_LIMITATION
            - OTHER
        reason_text:
          type: string
          minLength: 1
    RiskScoreAdjustment:
      type: object
      required: [score, decision, floor_override_warning]
      properties:
        score:
          $ref: '#/components/schemas/RiskScore'
        decision:
          $ref: '#/components/schemas/OperatorDecision'
        floor_override_warning:
          type: boolean
    OperatorDecision:
      type: object
      required:
        - decision_id
        - task_id
        - target_type
        - target_id
        - decision_type
        - after_json
        - reason_code
        - reason_text
        - operator_id
        - created_at
      properties:
        decision_id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        target_type:
          type: string
        target_id:
          type: string
          format: uuid
        decision_type:
          type: string
        before_json:
          type:
            - string
            - 'null'
        after_json:
          type: string
        reason_code:
          type: string
        reason_text:
          type: string
        operator_id:
          type: string
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      required:
        - code
        - message
        - trace_id
        - retryable
        - details
        - occurred_at
      properties:
        code:
          type: string
        message:
          type: string
        task_id:
          type:
            - string
            - 'null'
          format: uuid
        trace_id:
          type: string
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true
        occurred_at:
          type: string
          format: date-time